web_interface/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-socketio>=5.10.0
websockets>=12.0
sqlalchemy[asyncio]>=2.0.0
asyncmy>=0.2.8
alembic>=1.12.0
python-dateutil>=2.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
